requirements.txt
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
fastcore==1.7.19
fastlite==0.0.13
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.8.0
python-multipart==0.0.16
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==3.37.0.post4
starlette==0.41.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
